requirements_parser.txt
pdfminer.six==20181108
pycryptodome==3.7.3
six==1.12.0
sortedcontainers==2.1.0
wordsegment==1.3.1
python-docx==0.8.10
